requirements.txt
pytesseract>=0.3.10
Pillow>=10.3.0
